pyproject.toml
[project]
authors = [{name = "Cordero Core", email = "127983572+uwcdc@users.noreply.github.com"}]
dependencies = ["pyyaml", "requests", "beautifulsoup4"]
name = "wa-law-scraper"
requires-python = ">= 3.11"
version = "0.1.0"

[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[tool.hatch.build.targets.wheel]
packages = ["src/wa_law_scraper"]

[tool.pixi.workspace]
channels = ["https://prefix.dev/conda-forge"]
platforms = ["linux-64", "osx-arm64"]

[tool.pixi.feature.docs.dependencies]
mkdocs = "*"
mkdocs-material = "*"
mkdocs-include-markdown-plugin = "*"
mkdocstrings = "*"
mkdocstrings-python = "*"
jupyter-book = "*"
sphinx = "*"

[tool.pixi.pypi-dependencies]
wa_law_scraper = { path = ".", editable = true }

[tool.pixi.tasks]
docs-mkdocs-serve = "cd docs/mkdocs && mkdocs serve"
docs-mkdocs-build = "cd docs/mkdocs && mkdocs build"
docs-jupyter-book-build = "cd docs && jupyter-book build jupyter-book/"
docs-jupyter-book-serve = "cd docs && jupyter-book build jupyter-book/ --open"
docs-clean = "rm -rf docs/site docs/jupyter-book/_build"
# Registry CLI tasks
registry-generate-wac = "python -m wa_law_scraper.cli generate wac --verbose"
registry-generate-rcw = "python -m wa_law_scraper.cli generate rcw --verbose"
registry-generate-both = "python -m wa_law_scraper.cli generate both --verbose"
registry-generate-wac-rate-limited = "python -m wa_law_scraper.cli generate wac --rate-limit --verbose"
registry-generate-rcw-rate-limited = "python -m wa_law_scraper.cli generate rcw --rate-limit --verbose"
registry-generate-both-rate-limited = "python -m wa_law_scraper.cli generate both --rate-limit --verbose"
registry-list = "python -m wa_law_scraper.cli list"
registry-list-wac = "python -m wa_law_scraper.cli list --code-type wac"
registry-list-rcw = "python -m wa_law_scraper.cli list --code-type rcw"
registry-info-wac = "python -m wa_law_scraper.cli info --code-type wac"
registry-info-rcw = "python -m wa_law_scraper.cli info --code-type rcw"

[tool.pixi.dependencies]
pyyaml = ">=6.0.2,<7"
requests = ">=2.32.4,<3"
beautifulsoup4 = ">=4.13.4,<5"
fake-useragent = ">=2.2.0,<3"
